fix utils

+++ include/utils.hpp
@@ -3,7 +3,5 @@
 #include <string>
 #include <cstdint>
 
-
-void hex_to_bytes(const std::string& hex_string, uint64_t digits[], int& count);
-
-void bytes_to_hex(uint64_t A[], std::string& a, int count);
+void hex_to_bytes(const std::string& hex_string, uint8_t digits[], int& count);
+void bytes_to_hex(uint8_t A[], std::string& a, int count);
